Reviewed all off wendigo_scan.c with a defensive programming perspective - Eliminating any remaining opportunities to act on NULL values, work   outside of allocated memory, etc. Made a significant number of changes to parseBluetooth(), which was frequently (but not consistently) returning one too few bytes as consumed, leaving the final byte of the packet terminator in the buffer. I guess the changes to parseBuffer() to support fragmented or corrupted packets is working well, since that error was completely invisible :)
While making these changes I learned that strncpy doesn't copy or create
a null character to terminate the created string, which boggles the mind
a bit. From my understanding, if the specified size is less than the
size of the source string the destination will be padded with NULLs, but
where the lengths are consistent or strncpy() needs to truncate the
source string, it leaves you without a null terminator.
Makes you wonder why the function even exists since memcpy exists
*sigh*. Need to review all use of strncpy() to ensure I haven't relied
on this non-existent feature.